@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class FadeRouteBuilder<T> extends PageRouteBuilder<T> {
   final Widget page;
 
-  FadeRouteBuilder({@required this.page})
+  FadeRouteBuilder({required this.page})
       : super(
           pageBuilder: (context, animation1, animation2) => page,
           transitionsBuilder: (context, animation1, animation2, child) {
