@@ -5,7 +5,6 @@ class HotelConceptThemeProvider {
     return ThemeData(
       primaryColorLight: Color(0xff5b626b),
       hintColor: Color(0xffbfc2c5),
-      accentColor: Color(0xff006be3),
       unselectedWidgetColor: Color(0x191a86ff),
       indicatorColor: Color(0x33ffffff),
       highlightColor: Color(0xffe8f2ff),
@@ -28,6 +27,7 @@ class HotelConceptThemeProvider {
           headline1: TextStyle(
             color: Color(0xffededed),
           )),
+      colorScheme: ColorScheme.fromSwatch().copyWith(secondary: Color(0xff006be3)),
     );
   }
 }
